edge/requirements.txt
ultralytics>=8.0.0,<9.0.0
opencv-python-headless>=4.8.0,<5.0.0
paho-mqtt>=1.6.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
numpy>=1.24.0,<2.0.0
